+++ docx4j-export-fo/src/main/java/module-info.java
@@ -29,6 +29,7 @@ module org.docx4j.export_fo {
 	exports org.docx4j.convert.out.pdf.viaXSLFO;
 	exports org.docx4j.convert.out.XSLFO;
 	exports org.docx4j.fop.wordlayout;
+	exports org.docx4j.fop.fonts;
 
     // Resource folders must be open! See https://stackoverflow.com/questions/45166757/loading-classes-and-resources-in-java-9/45173837#45173837  
 		opens org.docx4j.convert.out.fo;